Refactor App data into its own file
